Yet anothe tide of changes, including the refactoring of asyncore and asynchat into Allegra, a final move to new style classes and out of lambda, an almost complete replacement of fifo by deque and many bugfixes all the way.
"Mille fois sur le métier remettez votre ouvrage"

It is so true. This is the third time I'm re-reading the library from
the beginning, and possibly the last. There is little to remove or add
from the first part (up and before the first Internet protocols). The
final test will of course come by polishing the implementation of
the PNS, HTTP, NNRP, SMTP and POP servers and/or clients.