Yet another see of changes. This revision includes the first module of beta-quality: netstring.py, simplified as much as possible and now with shorter and fewer interface names.
Also, I finally broke asyncore and asynchat into pieces, adding my
own async_net and paving the way for optimization (I mean, a C module
someday in the far future ...). Last but not least, this revision is
an allmost pure 2.4 implementation, with heavy use of deque, heapq,
sets and new-style classes, without a lambda, ready to climb the slope
up to the 3.0 version of CPython. 

Maybe broken mostly as far as the applications are concerned, but pretty
robust at the core! The support for DNS named TCP client has moved to
the dns_client module and I'm actively working on MIME, SMTP, NNRP and
HTTP clients, trying to figure out a stable interface. The next steps
down the road to a first Beta release will then be to integrate those
clients with PNS and Presto, allowing an Allegra peer to browse its
metabase and feed it with it's users mail, bookmarks and news feeds.