Yet another change to the asynchat interface and implementation. I added support for stallable collector in order to enable an HTTP server to defer instanciation of the MIME body collector for POST and PUT commands.
Also, I moved back the loop inside the functions that
collect and producer data from and to the buffers. Finally,
the readable and writable for stalled collector and producer
have replaced the originals. The rational for that default
behaviour is that Allegra is for peers not server or client
and that if you *need* something simpler (and faster) than
async_chat, the you should investigate netstring protocols
and use async_net instead.