Practicality beats purity ;-)
I have not really ever applied handle_expt_event, because
this amounts to actually checking and error condition
with select.select. And that is beyond my reach, deep into
systemland and the underlying socket implementation.

Doing without select's errors is practical and a tad faster.

Eventually it also allows to select more file descriptors
at once if many readable channels are not writable, or vice
versa.