One line more added to async_core.Dispatcher:
  make sure that socket == None in close ()

You don't want to inspect a dead socket once you handled
the error or close event, so it's safe to assume at close
that the socket is to be destroyed. Plus is helps finding
out if one has been created or not, and it prevents calls
to a socket that does not exist to report the error in the
application and not a socket error.